+++ concordium-consensus/src/Concordium/GlobalState/BlockState.hs
@@ -44,7 +44,7 @@ import Data.Serialize(Serialize)
 import qualified Data.Map as Map
 import qualified Data.Sequence as Seq
 import Data.Foldable (foldl')
-import qualified Data.ByteString.Lazy as LBS
+import qualified Data.ByteString as BS
 import Data.Word
 import System.IO (Handle)
 
@@ -256,7 +256,7 @@ class AccountOperations m => BlockStateQuery m where
     -- returns @Left protocolUpdate@. Otherwise, returns @Right pendingProtocolUpdates@.
     -- The @pendingProtocolUpdates@ is a (possibly-empty) list of timestamps and protocol
     -- updates that have not yet taken effect.
-    getProtocolUpdateStatus :: BlockState m -> m (Either ProtocolUpdate [(TransactionTime, ProtocolUpdate)])
+    getProtocolUpdateStatus :: BlockState m -> m UQ.ProtocolUpdateStatus
 
     -- |Get the current cryptographic parameters of the chain.
     getCryptographicParameters :: BlockState m -> m CryptographicParameters
@@ -526,6 +526,17 @@ class (BlockStateQuery m) => BlockStateOperations m where
   -- |Enqueue an update to take effect at the specified time.
   bsoEnqueueUpdate :: UpdatableBlockState m -> TransactionTime -> UpdateValue -> m (UpdatableBlockState m)
 
+  -- |Overwrite the election difficulty, removing any queued election difficulty updates.
+  -- This is intended to be used for protocol updates that affect the election difficulty in
+  -- tandem with the slot duration.
+  -- Note that this does not affect the next sequence number for election difficulty updates.
+  bsoOverwriteElectionDifficulty :: UpdatableBlockState m -> ElectionDifficulty -> m (UpdatableBlockState m)
+
+  -- |Clear the protocol update and any queued protocol updates.
+  -- This is intended to be used to reset things after a protocol update has taken effect.
+  -- This does not affect the next sequence number for protocol updates.
+  bsoClearProtocolUpdate :: UpdatableBlockState m -> m (UpdatableBlockState m)
+
   -- |Add the given accounts and timestamps to the per-block account release schedule.
   -- PRECONDITION: The given timestamp must be the first timestamp for a release for the given account.
   bsoAddReleaseSchedule :: UpdatableBlockState m -> [(AccountAddress, Timestamp)] -> m (UpdatableBlockState m)
@@ -594,7 +605,7 @@ class (BlockStateOperations m, Serialize (BlockStateRef m)) => BlockStateStorage
 
     -- |Serialize the block state to a byte string.
     -- This serialization does not include transaction outcomes.
-    serializeBlockState :: BlockState m -> m LBS.ByteString
+    serializeBlockState :: BlockState m -> m BS.ByteString
 
     -- |Serialize the block state to a file handle.
     -- This serialization does not include transaction outcomes.
@@ -711,6 +722,8 @@ instance (Monad (t m), MonadTrans t, BlockStateOperations m) => BlockStateOperat
   bsoGetUpdateKeyCollection = lift . bsoGetUpdateKeyCollection
   bsoGetNextUpdateSequenceNumber s = lift . bsoGetNextUpdateSequenceNumber s
   bsoEnqueueUpdate s tt payload = lift $ bsoEnqueueUpdate s tt payload
+  bsoOverwriteElectionDifficulty s = lift . bsoOverwriteElectionDifficulty s
+  bsoClearProtocolUpdate = lift . bsoClearProtocolUpdate
   bsoAddReleaseSchedule s l = lift $ bsoAddReleaseSchedule s l
   bsoGetEnergyRate = lift . bsoGetEnergyRate
   bsoGetChainParameters = lift . bsoGetChainParameters
@@ -754,6 +767,8 @@ instance (Monad (t m), MonadTrans t, BlockStateOperations m) => BlockStateOperat
   {-# INLINE bsoGetUpdateKeyCollection #-}
   {-# INLINE bsoGetNextUpdateSequenceNumber #-}
   {-# INLINE bsoEnqueueUpdate #-}
+  {-# INLINE bsoOverwriteElectionDifficulty #-}
+  {-# INLINE bsoClearProtocolUpdate #-}
   {-# INLINE bsoAddReleaseSchedule #-}
   {-# INLINE bsoGetEnergyRate #-}
   {-# INLINE bsoGetChainParameters #-}